backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
PyJWT==2.8.0
email-validator==2.1.0
Pillow==10.1.0
schedule==1.2.0
adata
pypinyin==0.49.0
APScheduler==3.10.4
